% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneDsample1.R
\name{oneDSample1}
\alias{oneDSample1}
\title{Single Variable Rejection Sampling}
\usage{
oneDSample1(f, N = 10000, lb = Inf, ub = Inf, continuous = TRUE)
}
\arguments{
\item{f}{the pdf that we are sampling from}

\item{N}{the number of attempted samples}

\item{lb}{the lower bound of support}

\item{ub}{the upper bound of support}
}
\value{
A vector containing samples from pdf
}
\description{
This function implements one variables rejection sampling for rvs
with bounded support and which have a bounded pdf.
}
\details{
Here are more details about the algorithm that we are using
}
\examples{
f<- function(x) {ifelse(-1< x & x < 0, 3*x^2, 0)}
w=oneDsample(f,50000,-1,0)
ggplot(w,aes(x)) + geom_density() + stat_function(fun = f, color = "red")

f<- function(x) {1/pi/(1+x^2)}
w=oneDSample1(f,50000)
ggplot(w,aes(x)) + geom_density() + stat_function(fun = f, color = "red")

f <- function(x) {dunif(x, -100, 100)}
a <- oneDSample1(f,500)
ggplot(a,aes(x)) + geom_density() + stat_function(fun = f, color = "red")


f<- function(x){dnorm(x,-5000,5000)}
w=oneDSample1(f,50000)
ggplot(w,aes(x)) + geom_density() + stat_function(fun = f, color = "red")

}
