% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoDSample.R
\name{twoDsample}
\alias{twoDsample}
\title{Double Variable Rejection Sampling}
\usage{
twoDsample(f, N = 1000, lbx = -1000, ubx = 1000, lby = -1000,
  uby = 1000)
}
\arguments{
\item{f}{the pdf that we are sampling from}

\item{N}{the number of attempted samples}

\item{lbx}{the lower bound of support x of f}

\item{ubx}{the upper bound of support x of f}

\item{lby}{the lower bound of support y of f}

\item{uby}{the upper bound of support y of f}
}
\value{
A vector containing samples from pdf
}
\description{
This function implements two variables rejection sampling for rvs
with bounded support and which have a bounded pdf.
}
\details{
Here are more details about the algorithm that we are using
}
\examples{


jointPFF <- function(x){
x1 = x[1]
x2 = x[2]
ifelse(0<x1 & x1<1 & 0<x2 & x2<1 , 24*x1*x2, 0)}
w = twoDsample(f = jointPFF, N=10000,0,1,0,1)
ggplot(w, aes(x, y)) +  geom_density_2d()

f <- function(x){
x1 = x[1]
x2 = x[2]
ifelse(x2>0, 1/pi/(1+x1^2) * 0.05*exp(-0.05*x2), 0)}
w=twoDsample(f = f, N=10000)
ggplot(a, aes(x, y)) +  geom_density_2d()
}
